Moved EventEntity into common module and fix fields name with snake case
